22860

See merge request akita/akita-erp!62

+++ public/assets/js/controller/suppliers-list.js
@@ -1150,11 +1150,34 @@ var ctrSuppliersListVl = new Vue({
     deleteSupplier: function deleteSupplier(id) {
       var _this2 = this;
 
-      if (confirm(messages["MSG06001"])) {
-        suppliers_service.deleteSupplier(id).then(function (response) {
+      suppliers_service.checkIsExist(id).then(function (response) {
+        if (!response.success) {
+          alert(response.msg);
+
           _this2.getItems(1);
-        });
-      }
+
+          return false;
+        } else {
+          if (confirm(messages["MSG06001"])) {
+            suppliers_service.deleteSupplier(id).then(function (response) {
+              _this2.getItems(1);
+            });
+          }
+        }
+      });
+    },
+    checkIsExist: function checkIsExist(id) {
+      var _this3 = this;
+
+      suppliers_service.checkIsExist(id).then(function (response) {
+        if (!response.success) {
+          alert(response.msg);
+
+          _this3.getItems(1);
+        } else {
+          window.location.href = 'edit/' + id;
+        }
+      });
     }
   },
   mounted: function mounted() {
